Bugs.

+++ kern/env.c
@@ -277,8 +277,8 @@ region_alloc(struct Env *e, void *va, size_t len)
 	if (e == NULL){
 		panic("region_alloc(): Point of struct Env e is NULL. Fail code: %e\n", -E_BAD_ENV);
 	}
-	uintptr_t lva = ROUNDDOWN(va, PGSIZE);
-	uintptr_t rva = ROUNDUP(va+len, PGSIZE);
+	uintptr_t lva = (unsigned int)ROUNDDOWN((unsigned int)va, PGSIZE);
+	uintptr_t rva = (unsigned int)ROUNDUP(va+len, PGSIZE);
 	struct PageInfo *p;
 	for( uintptr_t i = lva; i < rva; i+= PGSIZE){
 		p = page_alloc(0);
